requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
coverage==7.3.2
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
iniconfig==2.0.0
mypy==1.6.1
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
python-magic==0.4.27
requests==2.31.0
smmap==5.0.1
types-requests==2.31.0.10
typing_extensions==4.8.0
urllib3==2.0.7
